Fixed comparison between signed and unsigned integer expressions warnings.

+++ tsk/fs/fatfs_meta.c
@@ -1373,7 +1373,7 @@ fatfs_inode_walk(TSK_FS_INFO *a_fs, TSK_INUM_T a_start_inum,
             /* Read in a cluster. */
             cnt = tsk_fs_read_block
                 (a_fs, sect, dino_buf, num_sectors_to_process << fatfs->ssize_sh);
-            if (cnt != (num_sectors_to_process << fatfs->ssize_sh)) {
+            if (cnt != (ssize_t)(num_sectors_to_process << fatfs->ssize_sh)) {
                 if (cnt >= 0) {
                     tsk_error_reset();
                     tsk_error_set_errno(TSK_ERR_FS_READ);
